add claim letter service
