+++ src/Consumer/Taurus/Modules/TbClaimService.php
@@ -2,35 +2,10 @@
 
 namespace Taurus\Workflow\Consumer\Taurus\Modules;
 
-use App\Services\Claim\OnDemandWorkflowService;
-
 class TbClaimService extends ModuleService
 {
     public function getPostFixForTaskDefinition()
     {
         return 'claim';
-    }
-
-    public function getExtendedTemplateInfo(array $templatePayload = []): array
-    {
-        try {
-            $payload = [
-                'recordIdentifier' => getRecordIdentifierForRunningWorkflow() ?? null,
-                'workflowTemplateId' => $templatePayload['id'] ?? null,
-            ];
-
-            $response = OnDemandWorkflowService::fetchExtendedTemplateInfo($payload);
-
-            return $response;
-        } catch (\Exception $e) {
-            \Log::error('WORKFLOW - Error getting extended template info for TbClaim.', [
-                'message' => $e->getMessage(),
-                'trace' => $e->getTraceAsString(),
-                'line' => $e->getLine(),
-                'file' => $e->getFile(),
-            ]);
-
-            return [];
-        }
     }
 }
